Separate openloan logic into view and controller.
